+++ cynic/tests/simple_schema_tests.rs
@@ -6,7 +6,7 @@ mod query_dsl {
 
 use cynic::selection_set;
 
-#[derive(Clone, cynic::FragmentArguments)]
+#[derive(cynic::FragmentArguments)]
 struct TestArgs {}
 
 #[derive(cynic::QueryFragment, PartialEq, Debug)]
@@ -64,7 +64,7 @@ pub enum Dessert {
 fn run_test(input: serde_json::Value, expected_result: TestQuery) {
     use cynic::QueryFragment;
 
-    let query = cynic::Operation::query(TestQuery::fragment(TestArgs {}));
+    let query = cynic::Operation::query(TestQuery::fragment(&TestArgs {}));
 
     let test_data = cynic::GraphQLResponse {
         errors: None,